tui: fix a few lifetime bounds
Adding the ansi text container to TuiData seems to have moved the lifetimes around such that these are now required.. this is a little concerning (especially since the lifetime on AnsiTextContainer is basically unused by us), but things still seem to work.

+++ tui/src/tui/widget/util.rs
@@ -62,7 +62,7 @@ pub fn trim_to_width(inp: &str, width: u16) -> String {
     let width = width as usize;
     let num_chars = s.chars().count();
 
-    if num_chars <= width as usize {
+    if num_chars <= width {
         s
     } else {
         // We're assuming the width is _at least_ 2 here; otherwise this isn't
